+++ tests/python_api/audio_unit_test.py
@@ -39,7 +39,10 @@ test_case_augmentation_map = {
     1: "preemphasis_filter",
     2: "spectrogram",
     3: "downmix",
-    4: "to_decibels"
+    4: "to_decibels",
+    5: "resample",
+    6: "tensor_add_tensor",
+    7: "tensor_mul_scalar"
 }
 
 def plot_audio_wav(audio_tensor, idx):
@@ -138,6 +141,53 @@ def to_decibels_pipeline(path, file_list):
             cutoff_db=np.log(1e-20),
             output_dtype=types.FLOAT)
 
+@pipeline_def(seed=seed)
+def resample_pipeline(path, file_list):
+    audio, labels = fn.readers.file(file_root=path, file_list=file_list)
+    decoded_audio = fn.decoders.audio(
+        audio,
+        file_root=path,
+        file_list_path=file_list,
+        downmix=True,
+        shard_id=0,
+        num_shards=1,
+        stick_to_shard=False)
+    input_sample_rate = 16000.00
+    uniform_distribution_resample = fn.random.uniform(decoded_audio, range=[1.15, 1.15])
+    resampled_rate = uniform_distribution_resample * input_sample_rate
+    return fn.resample(
+        decoded_audio,
+        resample_rate=resampled_rate,
+        resample_hint=1.15 * 255840,
+        output_datatype=types.FLOAT)
+
+@pipeline_def(seed=seed)
+def tensor_add_tensor_pipeline(path, file_list):
+    audio, labels = fn.readers.file(file_root=path, file_list=file_list)
+    decoded_audio = fn.decoders.audio(
+        audio,
+        file_root=path,
+        file_list_path=file_list,
+        downmix=False,
+        shard_id=0,
+        num_shards=1,
+        stick_to_shard=False)
+    uniform_distribution_sample = fn.random.uniform(decoded_audio, range=[1.15, 1.15])
+    return decoded_audio + uniform_distribution_sample
+
+@pipeline_def(seed=seed)
+def tensor_mul_scalar_pipeline(path, file_list):
+    audio, labels = fn.readers.file(file_root=path, file_list=file_list)
+    decoded_audio = fn.decoders.audio(
+        audio,
+        file_root=path,
+        file_list_path=file_list,
+        downmix=True,
+        shard_id=0,
+        num_shards=1,
+        stick_to_shard=False)
+    return decoded_audio * 1.15
+
 def main():
     args = parse_args()
 
@@ -201,6 +251,12 @@ def main():
                                                     path=downmix_audio_path if qa_mode else audio_path, file_list="" if qa_mode else file_list, downmix=True)
         if case_name == "to_decibels":
             audio_pipeline = to_decibels_pipeline(batch_size=batch_size, num_threads=num_threads, device_id=device_id, rocal_cpu=rocal_cpu, path=audio_path, file_list=file_list)
+        if case_name == "resample":
+            audio_pipeline = resample_pipeline(batch_size=batch_size, num_threads=num_threads, device_id=device_id, rocal_cpu=rocal_cpu, path=audio_path, file_list=file_list)
+        if case_name == "tensor_add_tensor":
+            audio_pipeline = tensor_add_tensor_pipeline(batch_size=batch_size, num_threads=num_threads, device_id=device_id, rocal_cpu=rocal_cpu, path=audio_path, file_list=file_list)
+        if case_name == "tensor_mul_scalar":
+            audio_pipeline = tensor_mul_scalar_pipeline(batch_size=batch_size, num_threads=num_threads, device_id=device_id, rocal_cpu=rocal_cpu, path=audio_path, file_list=file_list)
         audio_pipeline.build()
         audio_loader = ROCALAudioIterator(audio_pipeline, auto_reset=True)
         output_tensor_list = audio_pipeline.get_output_tensors()