+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 * Support for Audio augmentation - Spectrogram
 * Support for Audio augmentation - ToDecibels
 * Support for downmixing audio channels during decoding
+* Support for Audio augmentation - Resample
+* Support for TensorTensorAdd and TensorScalarMultiply operations
+* Support for Uniform and Normal distribution nodes
 
 ### Optimizations